WebProject/requirements.txt
asgiref==3.2.5
Django==3.0.4
django-taggit==1.2.0
Faker==4.0.2
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uWSGI==2.0.18
